Trying to fix negative instants

+++ global.h
@@ -22,7 +22,7 @@
 #define WRITE 1
 
 
-long int start; //Start time in milisseconds for logs
+double start; //Start time in milisseconds for logs
 
 //Arrays used for logs
 char log_filename[STR_LEN];
